refactor/파일 및 클래스 구조 변경
