--- 이우성/_1/Student.java
@@ -1,27 +1,0 @@
-package _1;
-
-public class Student {
-	
-	private String name;
-	private String no;
-	
-	public Student(String name, String no) {
-		super();
-		this.name = name;
-		this.no = no;
-	}
-	
-	public String getName() {
-		return name;
-	}
-	public void setName(String name) {
-		this.name = name;
-	}
-	public String getNo() {
-		return no;
-	}
-	public void setNo(String no) {
-		this.no = no;
-	}
-
-}
